Removed "setrule()" in the constructor of generations.
Also made the rule survive alloc/frees.  We need to move the rule in
general to algoData where it belongs (a safe static place to keep it
that survives universe construction/destruction) and make this
somewhat automatic.  Indeed, each lifealgo should have direct
access to its own algodata.  More on that as we make progress.