Fixed one bug in qlifealgo (just a workarond).  Every time editing was done, setrule() was called, which called markglobalchange(), which caused an uproot().  Eventually these uproots caused qlifealgo to grow too deep.
This also checks in the first support for tracking changes to
global_liferule in global_liferule itself, and allowing algoirthms to pick
up on that change as they need to.  This way we can support
multiple live algorithms at the same time but still have them share
a global liferule object.  We need to move the check of the serial
value from setrule() to dostep(), however, and I'm not quite ready
to make that change yet.