Refactor GTK Makefile
This commit refactors the makefiles with the following changes:

1. Extract program names as variables.
 This includes WX_CONFIG and PYTHON. The rationale of this change is:
it now allows to use different versions of the same program.
For instance, now one could use wxWidgets 2.8 and 3.1 on the same
system (on different directories), and specify which wx-config to use.
Same applies to Python.

2. Extract user-specific directories as variables.
This includes IRRKLANGDIR.

3. Add EXTRACXXFLAGS, EXTRALDFLAGS as in local-win-template.
This can be useful whenever we need extra flags conditionally.
For instance, we sometimes need rpath to be relative to the build,
so we add the appropriate flags here.

4. Minor changes ("Just in case" fixes)
- When loading Python compile and link flags, print(X) is fine when
printing one argument (parsed as print (X)) and is futureproof.
- Changed BITSIZE to `getconf LONG_BIT`, just in case this script gets
more portable. (shell seems to be a GNU make extension)

Unfortunately this involves "import" directives, and I am not sure how
portable this is... (at least works in GNU make)