Add support in HashLife algorithm for reading multi-state macrocell format.
Feature request: https://sourceforge.net/p/golly/support-requests/8/

This allows a macrocell file saved e.g. by the RuleTable algorithm to be
loaded by HashLife, provided that only binary (0 or 1) cell states are used.

This works by loading nodes that represent 2-squares and 4-squares into a leaf
cell of size 8x8, centered in the middle. When a leaf cell is used as the root
of a timeline savepoint, it is converted to an internal node, which is necessary
because the HashLife algorithm does not support a leaf node to be used as the
root.