Fix for Android 15
Starting in Android 15 (API 35), the action bar for the app overlaps
into the app itself and obscures the top bar with "Start", "Step",
"Control", etc.

This can be seen in this user report:
https://conwaylife.com/forums/viewtopic.php?f=4&p=202231#p202228

The cause is that starting in Android 15, Google has started enforcing
"edge-to-edge":
https://developer.android.com/about/versions/15/behavior-changes-15#edge-to-edge

The clue to fixing is the 3rd bullet point here:
https://developer.android.com/about/versions/15/behavior-changes-15#not-edge-to-edge

Adding android:fitsSystemWindows="true" to each of the activities fixes
the overlap problem.

The documentation for android:fitsSystemWindows says:
Boolean internal attribute to adjust view layout based on system
windows such as the status bar. If true, adjusts the padding of this
view to leave space for the system windows. Will only take effect if
this view is in a non-embedded activity.